Divide each layer's Gram matrix by height*width of the feature map
According to the paper, the style loss is El = 1/(4 * Nl^2 * Ml^2) * sum((Gl-Al)^2), where Nl is layer depth and Ml = width*height. Calculating the mean instead of sum eliminates the need to divide by Nl^2, but we still have to divide by Ml^2. And it's easier to do that when building a style map, because at the moment of calculating the loss we don't know what size of the feature map had been before we computed the Gram matrix.

https://github.com/udacity/deep-learning-v2-pytorch/issues/174